1530 Add confirmation and name modal
