+++ plugins/chunter-resources/src/utils.ts
@@ -39,16 +39,14 @@ export function classIcon (client: Client, _class: Ref<Class<Obj>>): Asset | und
   return client.getHierarchy().getClass(_class).icon
 }
 
-export async function getDmName (client: Client, dm: Space, shouldIncludeMyAcc = false): Promise<string> {
+export async function getDmName (client: Client, dm: Space): Promise<string> {
   const myAccId = getCurrentAccount()._id
 
   const employeeAccounts = await client.findAll(contact.class.EmployeeAccount, {
     _id: { $in: dm.members as Array<Ref<EmployeeAccount>> }
   })
 
-  const name = (
-    dm.members.length > 1 && !shouldIncludeMyAcc ? employeeAccounts.filter((a) => a._id !== myAccId) : employeeAccounts
-  )
+  const name = (dm.members.length > 1 ? employeeAccounts.filter((a) => a._id !== myAccId) : employeeAccounts)
     .map((a) => formatName(a.name))
     .join(', ')
